requirements.txt
# Core Web App
streamlit
fastapi
uvicorn

# Embedding and NLP
sentence-transformers
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5

# Text Search & Vector Databases
faiss-cpu
pinecone  # ✅ Correct package name (not pinecone-client)
pinecone-client>=3.0.0,<4.0.0
# PDF & Parsing
PyMuPDF

# Google Generative AI
google-generativeai
protobuf  # Needed for Google GenAI

# Utilities
requests
pydantic
python-dotenv 
